src/requirements.txt
setuptools
flask
flask_sqlalchemy
bcrypt
flake8
mypy
pytest
pytest-flask
python-dotenv
psycopg2-binary